requirements.txt
Django==4.1.5
python_dateutil==2.8.2
requests==2.28.2
tailwind
django_browser_reload
django_tailwind